fix(app): fix user crud routes

+++ src/infra/database/tables/card/card-table.ts
@@ -4,7 +4,7 @@ export const cardTable: string = `
         id VARCHAR(50) PRIMARY KEY NOT NULL,
         name VARCHAR(100) NOT NULL,
         nickname VARCHAR(100) NOT NULL,
-        number INTEGER NOT NULL,
+        number NUMERIC NOT NULL,
         securityCode INTEGER NOT NULL,
         createdAt VARCHAR(10) NOT NULL,
         updatedAt VARCHAR(10) NOT NULL
